- 통계 일괄 수정

+++ src/main/java/com/ads/main/vo/adGroup/resp/PartnerAdGroupVo.java
@@ -87,7 +87,7 @@ public class PartnerAdGroupVo implements Serializable {
     /**
      * 포인트 교환 비율
      */
-    private Integer rewordRate;
+    private Double rewordRate;
 
 
     /**
